requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.3.0
backoff==2.2.1
bcrypt==4.2.1
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.15
distro==1.9.0
durationpy==0.9
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
filelock==3.16.1
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.12.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==31.0.0
langchain==0.3.14
langchain-chroma==0.2.0
langchain-community==0.3.14
langchain-core==0.3.29
langchain-ollama==0.2.2
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langgraph==0.2.62
langgraph-checkpoint==2.0.9
langgraph-cli==0.1.67
langgraph-sdk==0.1.51
langsmith==0.2.10
markdown-it-py==3.0.0
marshmallow==3.25.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.6
onnxruntime==1.20.1
openai==1.59.7
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.14
overrides==7.7.0
packaging==24.2
parso==0.8.4
posthog==3.8.3
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
